ffi: move initconfig.rs to cpython/initconfig.rs
