+++ src/ffi/cpython/mod.rs
@@ -9,10 +9,13 @@ pub mod dictobject;
 // skipped fileobject.h
 pub mod frameobject;
 // skipped import.h
-// skipped initconfig.h
+#[cfg(all(Py_3_8, not(PyPy)))]
+pub mod initconfig;
 // skipped interpreteridobject.h
 pub mod listobject;
 pub mod object;
+#[cfg(all(Py_3_8, not(PyPy)))]
+pub mod pylifecycle;
 
 pub use self::abstract_::*;
 #[cfg(not(PyPy))]
@@ -22,5 +25,9 @@ pub use self::code::*;
 #[cfg(not(PyPy))]
 pub use self::dictobject::*;
 pub use self::frameobject::*;
+#[cfg(all(Py_3_8, not(PyPy)))]
+pub use self::initconfig::*;
 pub use self::listobject::*;
 pub use self::object::*;
+#[cfg(all(Py_3_8, not(PyPy)))]
+pub use self::pylifecycle::*;
